remove lib.rs in module
